Got camera working on car. Started filtering (1,2,4,2,1) working but derivative not working. Got a standard turn radius.

+++ SRC/pwm.c
@@ -55,17 +55,20 @@ void SetMotorDutyCycle(unsigned int DutyCycle, unsigned int Frequency, int dir)
 
 
 
-/*
- * Change the Servo Duty Cycle and Frequency
- * @param DutyCycle (0 to 100)
- */
-void SetServoDutyCycle(unsigned int DutyCycle)
+/* SetServoDutyCycle
+* Description:
+*	Change the Servo Duty Cycle and Frequency
+*
+* Parameters:
+* 	DutyCycle - 4.6 to 9.14545 are full left and full right respectively.
+*/
+void SetServoDutyCycle(double DutyCycle)
 {
     // Calculate the new cutoff value
     uint16_t mod = (uint16_t) (((CLOCK/128/SERVO_FREQUENCY) * DutyCycle) / 100);
 
     // Set outputs
-  FTM3_C4V = mod;
+	FTM3_C4V = mod;
 
     // Update the clock to the new frequency
     FTM3_MOD = FTM3_MOD_VALUE;
@@ -79,7 +82,7 @@ void InitPWM()
 {
     // 12.2.13 Enable the Clock to the FTM0 Module
     SIM_SCGC6 |= SIM_SCGC6_FTM0_MASK;
-  SIM_SCGC3 |= SIM_SCGC3_FTM3_MASK;
+	SIM_SCGC3 |= SIM_SCGC3_FTM3_MASK;
 
     // Enable clock on PORT A so it can output
     SIM_SCGC5 |= SIM_SCGC5_PORTA_MASK | SIM_SCGC5_PORTB_MASK | SIM_SCGC5_PORTC_MASK;
@@ -113,7 +116,7 @@ void InitPWM()
 
     // 39.3.5 Set the Modulo resister
     FTM0_MOD = FTM0_MOD_VALUE;
-  FTM3_MOD = FTM3_MOD_VALUE;
+	FTM3_MOD = FTM3_MOD_VALUE;
 
     // 39.3.6 Set the Status and Control of both channels
     // Used to configure mode, edge and level selection
@@ -149,17 +152,17 @@ void InitPWM()
     // Chose system clock source
     // Timer Overflow Interrupt Enable
     FTM0_SC = FTM_SC_PS(0) | FTM_SC_CLKS(1);
-  FTM3_SC = FTM_SC_PS(7) | FTM_SC_CLKS(1);
+	FTM3_SC = FTM_SC_PS(7) | FTM_SC_CLKS(1);
 
     // Enable Interrupt Vector for FTM
-  //NVIC_EnableIRQ(FTM0_IRQn);
-  //NVIC_EnableIRQ(FTM3_IRQn);
+	//NVIC_EnableIRQ(FTM0_IRQn);
+	//NVIC_EnableIRQ(FTM3_IRQn);
   
-  // Set PTB2 and PTB3 for GPIO
+	// Set PTB2 and PTB3 for GPIO
     PORTB_PCR3 |= PORT_PCR_MUX(1);
     PORTB_PCR2 |= PORT_PCR_MUX(1);
   
-  // Set GPIO to output mode
+	// Set GPIO to output mode
     GPIOB_PDDR |= (1<<2) | (1<<3);
     
     // Enable Motors
